% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGradmatch.R
\docType{package}
\name{shinyGradmatch}
\alias{shinyGradmatch}
\alias{shinyGradmatch-package}
\title{shinyGradmatch: A package for computating the notorious bar statistic.}
\description{
The shinyGradmatch package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

