% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyKGode.R
\docType{package}
\name{shinyKGode}
\alias{shinyKGode}
\alias{shinyKGode-package}
\title{shinyKGode: A Shiny package for gradient matching parameter inference on 
ordinary differential equations.}
\description{
The shinyKGode package provides three main functions:
a friendly Shiny-based user-interface to run gradient matching to solve 
ordinary differential equations, a selection of three built-in models
and also the ability to import user-defined differential equations and 
perform parameter inference on them.
}
