% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{startShinyGradmatch}
\alias{startShinyGradmatch}
\title{This function will start ShinyGradmatch}
\usage{
startShinyGradmatch()
}
\value{
Nothing
}
\description{
An interactive Shiny application for running gradient matching to approximate ODEs.
}
\details{
Start ShinyGradmatch

This starts the ShinyGradmatch application on the users local computer.
}
\examples{
\dontrun{
startShinyGradmatch()
}
}
\keyword{ShinyGradmatch}
