% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{startShinyKGode}
\alias{startShinyKGode}
\title{This function will start ShinyKGode.}
\usage{
startShinyKGode()
}
\value{
Nothing
}
\description{
An interactive Shiny application for running gradient matching to approximate ODEs.
}
\details{
Start ShinyKGode

This starts the ShinyKGode application on the user's local computer.
}
\examples{
{
if(interactive()){
  startShinyKGode()
  }
}
}
\keyword{ShinyKGode}
